#8 - Implement cpu.hpp
