+++ src/calculator.hpp
@@ -1,25 +1,5 @@
 #include <interfaces.hpp>
 
-class Cpu : CpuInterface
-{
-    float *memory;
-    Operation operation;
-    unsigned short writeIndex;
-
-public:
-    Cpu();
-
-    ~Cpu();
-
-    float calculate();
-
-    void ac();
-
-    void write(float value);
-
-    void setOperation(Operation value);
-};
-
 class Display : DisplayInterface
 {
 public:
